fix(confenge): persist auditable route metadata
